Boot Mode normal and Boot status is abnormal, processing
Former-commit-id: 5f00cbbfdbb48d8241a6a4d511ea8d78179dc7a3

+++ SSBLFactory/SBC_BootProc.h
@@ -22,7 +22,7 @@ typedef struct _boot_proc_t {
     UINTN   curr_sw_bnk;            /*! Current SW Bank ID */
     UINT16  bm;                     /*! Boot Mode */
     UINT16  km;                     /*! Key Mode */
-    UINT8   bootst;                 /*! Boot Status */
+    UINT32   bootst;                 /*! Boot Status */
 }boot_proc_t;
 #pragma pack()
